Refactoring the Balchivist library.
This change moves many of the common functions used in the current
dumps script into the library and isolates the dumps archiving
script into an individual module inside the modules/ directory.
This change will make it easier for new modules to be created as
there is an easily-reusable set of functions available for all
modules to use.

Internationalization is now possible, though it is currently not
used due to the lack of support for it in the Internet Archive.